+++ pysyringe/__init__.py
@@ -1,1 +1,5 @@
-__version__ = "1.5.3.dev3+g25a74e36c"
+__version__ = "2.0.0"
+
+from pysyringe.container import Container, Provide
+
+__all__ = ["Container", "Provide"]